Add option to enable data checksums (postgresql_data_checksums)
The option is turned on by default, whih is a best practice
because it safeguards data integrity at a minimal performance
cost.

"Use checksums on data pages to help detect corruption by the I/O system
that would otherwise be silent. Enabling checksums may incur a
noticeable performance penalty. This option can only be set during
initialization, and cannot be changed later. If set, checksums are
calculated for all objects, in all databases."