Reload or restart pg if conf changed
On non-prd environments we can accept to restart postgresql in all
cases if configuration has changed.

In a better and future version, we could split and improve the reload
or restart task in order to:
- continue to never restart on prd through the handler
- on non-prd, restart depending on the changed file and parameter
(e.g: pg_hba.conf file doesn't require restart, and only some
postgresql.conf parameters require a restart).